requirements.txt
streamlit
pandas
gspread
google-auth-oauthlib
google-api-python-client
oauth2client
google-auth-httplib2